+++ src/components/ModalContent/ModalContent.js
@@ -18,16 +18,30 @@ import { stripesConnect } from '@folio/stripes/core';
 
 import { getOpenRequestsPath } from '../util';
 
-import { loanActionMutators, refundClaimReturned, loanActions } from '../../constants';
+import {
+  loanActionMutators,
+  refundClaimReturned,
+  loanActions,
+  CIRCULATION_BFF_LOANS_INTERFACE_NAME,
+  CIRCULATION_BFF_LOANS_INTERFACE_VERSION,
+  CIRCULATION_BFF_LOANS_INTERFACE_ERROR,
+} from '../../constants';
 
 import css from './ModalContent.css';
 
 export const getMutatorFunction = (stripes, mutator, loanAction) => {
-  if (stripes?.config?.enableEcsRequests && loanAction === loanActionMutators.DECLARE_LOST) {
-    if (stripes.hasInterface('circulation-bff-loans', '1.4')) {
+  const isDeclareLostAction = loanAction === loanActionMutators.DECLARE_LOST;
+  const isClaimedReturnedAction = loanAction === loanActionMutators.CLAIMED_RETURNED;
+
+  if (stripes?.config?.enableEcsRequests && (isDeclareLostAction || isClaimedReturnedAction)) {
+    const isCorrectCirculationBFFLoansInterface = stripes.hasInterface(CIRCULATION_BFF_LOANS_INTERFACE_NAME, CIRCULATION_BFF_LOANS_INTERFACE_VERSION);
+
+    if (isCorrectCirculationBFFLoansInterface && isDeclareLostAction) {
       return mutator.declareLostBFF.POST;
+    } else if (isCorrectCirculationBFFLoansInterface && isClaimedReturnedAction) {
+      return mutator.claimReturnedBFF.POST;
     } else {
-      throw new Error('Required okapi interfaces circulation-bff-loans v1.4');
+      throw new Error(CIRCULATION_BFF_LOANS_INTERFACE_ERROR);
     }
   } else {
     return mutator[loanAction].POST;
@@ -57,6 +71,14 @@ class ModalContent extends React.Component {
       throwErrors: false,
       POST: {
         path: 'circulation-bff/loans/!{loan.id}/declare-item-lost',
+      },
+    },
+    claimReturnedBFF: {
+      type: 'okapi',
+      fetch: false,
+      throwErrors: false,
+      POST: {
+        path: 'circulation-bff/loans/!{loan.id}/claim-item-returned',
       },
     },
     markAsMissing: {
@@ -102,6 +124,12 @@ class ModalContent extends React.Component {
         POST: PropTypes.func.isRequired,
       }).isRequired,
       declareLost: PropTypes.shape({
+        POST: PropTypes.func.isRequired,
+      }).isRequired,
+      declareLostBFF: PropTypes.shape({
+        POST: PropTypes.func.isRequired,
+      }).isRequired,
+      claimReturnedBFF: PropTypes.shape({
         POST: PropTypes.func.isRequired,
       }).isRequired,
       markAsMissing: PropTypes.shape({